requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain-community
pypdf
sentence-transformers
langchain-huggingface
faiss-cpu
scikit-learn
pydantic